refactor!: 删除命令行版，项目只剩图形界面
图形界面版功能已是命令行版的超集，命令行版不再需要。

删掉：
- src/gunlab.cpp（命令行界面层）
- scripts/build_msvc.bat、scripts/build_gcc.bat（都是构建 CLI 的）
- CMakeLists 里的 gunlab 目标
- zh_cn.h 里的 zh::t 命名空间（CLI 专用文案，215 行）和 zh::hit_tier()
  （长格式档位名，只有 CLI 用它；界面用的是 hit_tier_short）
- gunlab_math.cpp 里指向 hit_tier 的注释

★ 连带后果：MSVC 构建能力一并消失。GUI 需要 SDL3，而本机只有 msys64 的
  MinGW 版，MSVC 链不了。项目从此只支持 MinGW。

CMakeLists 里 find_package(SDL3) 由「找不到就跳过 GUI 目标」改成
FATAL_ERROR —— 以前跳过是因为还有命令行版可编，现在没有退路了，
早报错比晚报错好。报错信息里写了怎么装 SDL3、以及别用 MSVC 生成器。

CLAUDE.md 相应重写：构建只剩一种方式（build_gui.bat）、验证数值改用
「临时小程序链接 gunlab_core」（比原来的命令行版更好用：能按 id 定位枪、
能循环跑一批用例、能只打印关心的项），并补上「手工 cmake 重建后必须
跑 copy_gui_deps.ps1 拷 DLL」——我这次就漏了，现象是截图脚本报
「没拿到窗口句柄」，看不出真正原因。

验证：全量重建（删掉 build-gui 重新 configure）0 警告 0 错误；
补齐 DLL 后截图确认界面正常，所有数值与删除前逐项一致。

+++ src/gui/main_gui.cpp
@@ -1234,9 +1234,13 @@ void draw_instance( const DetailCache &d )
             const int rng = range_with_even_chance_of_good_hit( total );
             ImGui::TableNextRow();
             ImGui::TableNextColumn(); ImGui::TextUnformatted( r.name );
-            ImGui::TableNextColumn(); ImGui::Text( "%.0f", r.recoil );
-            ImGui::TableNextColumn(); ImGui::Text( "%.0f", d.fixed_disp );
-            ImGui::TableNextColumn(); ImGui::Text( "%.0f", total );
+            // ★ 一律用 (int) 截断，不用 %.0f。这些值都是小数（普通档阈值是
+            //   348.6、固定散布 299.5），四舍五入会显示成 349 / 300，和同屏
+            //   的图例（用截断）差一个数，看着像 bug。游戏自己也是
+            //   static_cast<int>（0.I ranged.cpp:2221）。
+            ImGui::TableNextColumn(); ImGui::Text( "%d", (int)r.recoil );
+            ImGui::TableNextColumn(); ImGui::Text( "%d", (int)d.fixed_disp );
+            ImGui::TableNextColumn(); ImGui::Text( "%d", (int)total );
             ImGui::TableNextColumn();
             if( rng >= 59 ) ImGui::TextUnformatted( zh::g::OVER_TABLE );
             else            ImGui::Text( zh::g::TILE_FMT, rng );
